fixed issues with dependency cycle and updated Readme file
