+++ README.md
@@ -1,16 +1,19 @@
-# Project Name
+# Weather App
 
-> One paragraph statement about the project.
+> Weather App fetches data from Open Weather Map API.
 
-![screenshot](./app_screenshot.png)
+![screenshot](./dist/images/home.png)
+![screenshot](./dist/images/error.png)
+![screenshot](./dist/images/results.png)
 
-Additional description about the project and its features.
+This projects implements using fetch method to return data from the weather api and return promises. The project also uses the try-catch block to handle scenarios where there is an error in the response from the API
 
 ## Built With
 
-- Major languages,
-- frameworks,
-- technologies used
+- HTML,
+- CSS,
+- JavaScript,
+- Webpack
 
 ## Live Demo
 
@@ -19,55 +22,46 @@ Additional description about the project and its features.
 
 ## Getting Started
 
-**This is an example of how you may give instructions on setting up your project locally.**
-**Modify this file to match your project, remove sections that don't apply. For example: delete the testing section if the currect project doesn't require testing.**
-
-
 To get a local copy up and running follow these simple example steps.
 
 ### Prerequisites
 
+- Have latest version of JS installed.
+- Have Node and NPM installed.
+- Install the latest version of your browser of choice.
+
 ### Setup
 
-### Install
+- Fork the repo to your remote repository.
+- Clone or download the repository to your local machine.
 
 ### Usage
 
-### Run tests
-
-### Deployment
-
-
+- Run npm installto install the project pakages
+- Run npm run devto bundle the packages for development. Dev is created for development, meaning that it supports hot module replacement, dev server, and a lot of things that assist your dev work
+- Run npm run build when you want to deploy the project for production.
+- Open the index.htmlfile in your browser.
+- Type in a correct city or town name in the input field.
+- If the input value is correct, the weather information about that city or town will be displayed.
+- Feel free to change the temperature from fahrenheit to celcius and vice versa by clicking the Temperature toggle Button
 
 ## Authors
 
-👤 **Author1**
+👤 **Rajkumar Das**
 
-- Github: [@githubhandle](https://github.com/githubhandle)
-- Twitter: [@twitterhandle](https://twitter.com/twitterhandle)
-- Linkedin: [linkedin](https://linkedin.com/linkedinhandle)
-
-👤 **Author2**
-
-- Github: [@githubhandle](https://github.com/githubhandle)
-- Twitter: [@twitterhandle](https://twitter.com/twitterhandle)
-- Linkedin: [linkedin](https://linkedin.com/linkedinhandle)
+- Github: [@rajkumardas2701](https://github.com/rajkumardas2701)
+- Twitter: [@Rajkuma58621299](https://twitter.com/Rajkuma58621299)
+- Linkedin: [Rajkumar Das](https://www.linkedin.com/in/rajkumar-das-41308961/)
 
 ## 🤝 Contributing
 
 Contributions, issues and feature requests are welcome!
 
-Feel free to check the [issues page](issues/).
+Feel free to check the [issues page](https://github.com/rajkumardas2701/WeatherApp_JS/issues).
 
 ## Show your support
 
 Give a ⭐️ if you like this project!
-
-## Acknowledgments
-
-- Hat tip to anyone whose code was used
-- Inspiration
-- etc
 
 ## 📝 License
