+++ src/appLogic.js
@@ -1,12 +1,12 @@
-import report from './weatherData';
+// import report from './weatherData';
 
-const controller = () => {
-  document.getElementById('form-submit').addEventListener('click', (event) => {
-    event.preventDefault();
-    const cityName = document.getElementById('form-text').value;
-    report(cityName);
-  });
-};
+// const controller = () => {
+//   document.getElementById('form-submit').addEventListener('click', (event) => {
+//     event.preventDefault();
+//     const cityName = document.getElementById('form-text').value;
+//     report(cityName);
+//   });
+// };
 
 const togglebtn = (element = '', temp = '') => {
   let currentUnit = '';
@@ -26,4 +26,4 @@ const togglebtn = (element = '', temp = '') => {
   }
 };
 
-export { controller, togglebtn };
+export default togglebtn;
