+++ src/weatherData.js
@@ -20,4 +20,12 @@ const report = async (city) => {
   }
 };
 
-export default report;
+const controller = () => {
+  document.getElementById('form-submit').addEventListener('click', (event) => {
+    event.preventDefault();
+    const cityName = document.getElementById('form-text').value;
+    report(cityName);
+  });
+};
+
+export default controller;
